refactor(core): improve flexibility of build, allow different touch panel driver, allow STM32F429
[no changelog]
