+++ core/embed/reflash/main.c
@@ -31,7 +31,7 @@
 #include "sbu.h"
 #include "sdcard.h"
 #include "secbool.h"
-#include "touch.h"
+#include "touch/touch.h"
 
 static void progress_callback(int pos, int len) { display_printf("."); }
