+++ core/embed/trezorhal/boards/trezor_t.h
@@ -1,6 +1,10 @@
 #ifndef _TREZOR_T_H
 #define _TREZOR_T_H
 
+#define USE_SD_CARD 1
+#define USE_TOUCH 1
+#define USE_SBU 1
+
 #include "displays/st7789v.h"
 
 #endif  //_TREZOR_T_H
