+++ core/embed/trezorhal/stm32f4xx_hal_conf.h
@@ -51,6 +51,10 @@
 #ifdef STM32F427xx
 #define USE_USB_HS
 #define USE_USB_HS_IN_FS
+#elif STM32F429xx
+#define USE_USB_HS
+#define USE_USB_HS_IN_FS
+#define HAL_LTDC_MODULE_ENABLED
 #elif STM32F405xx
 #define USE_USB_FS
 #else
