+++ core/embed/trezorhal/touch/ft6x36.c
@@ -24,6 +24,7 @@
 #include "common.h"
 #include "secbool.h"
 
+#include "ft6x36.h"
 #include "touch.h"
 
 #define TOUCH_ADDRESS \
@@ -359,22 +360,4 @@ uint32_t touch_read(void) {
   }
 
   return 0;
-}
-
-uint32_t touch_click(void) {
-  uint32_t r = 0;
-  // flush touch events if any
-  while (touch_read()) {
-  }
-  // wait for TOUCH_START
-  while ((touch_read() & TOUCH_START) == 0) {
-  }
-  // wait for TOUCH_END
-  while (((r = touch_read()) & TOUCH_END) == 0) {
-  }
-  // flush touch events if any
-  while (touch_read()) {
-  }
-  // return last touch coordinate
-  return r;
 }
