+++ core/embed/unix/touch/touch.h
@@ -1,0 +1,1 @@
+../../trezorhal/touch/touch.h
