+++ core/site_scons/boards/__init__.py
@@ -1,0 +1,3 @@
+def get_hw_model_as_number(hw_model):
+    return int.from_bytes(hw_model.encode(), 'little')
+
