+++ core/site_scons/tools.py
@@ -1,6 +1,8 @@
 import os.path
 
 import subprocess
+from boards import trezor_1, trezor_r_v3, trezor_r_v4, trezor_t
+
 
 
 def add_font(font_name, font, defines, sources):
@@ -14,40 +16,20 @@ def add_font(font_name, font, defines, sources):
             sources.append(sourcefile)
 
 
-def get_hw_model_as_number(hw_model):
-    return int.from_bytes(hw_model.encode(), 'little')
-
-
-def configure_board(model, env, defines, sources):
+def configure_board(model, features_wanted, env, defines, sources):
     model_r_version = 4
 
     if model in ('1',):
-        board = 'trezor_1.h'
-        display = 'vg-2864ksweg01.c'
-        hw_model = get_hw_model_as_number('T1B1')
-        hw_revision = 0
+        return trezor_1.configure(env, features_wanted, defines, sources)
     elif model in ('T',):
-        board = 'trezor_t.h'
-        display = 'st7789v.c'
-        hw_model = get_hw_model_as_number('T2T1')
-        hw_revision = 0
+        return trezor_t.configure(env, features_wanted, defines, sources)
     elif model in ('R',):
-        hw_model = get_hw_model_as_number('T2B1')
-        hw_revision = model_r_version
         if model_r_version == 3:
-            board = 'trezor_r_v3.h'
-            display = "ug-2828tswig01.c"
+            return trezor_r_v3.configure(env, features_wanted, defines, sources)
         else:
-            board = 'trezor_r_v4.h'
-            display = 'vg-2864ksweg01.c'
+            return trezor_r_v4.configure(env, features_wanted, defines, sources)
     else:
         raise Exception("Unknown model")
-
-    defines += [f'TREZOR_BOARD=\\"boards/{board}\\"', ]
-    defines += [f'HW_MODEL={hw_model}', ]
-    defines += [f'HW_REVISION={hw_revision}', ]
-    sources += [f'embed/trezorhal/displays/{display}', ]
-    env.get('ENV')['TREZOR_BOARD'] = board
 
 
 def get_model_identifier(model):
